requirements.txt
earthaccess==0.9.0
xarray==2024.1.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
requests==2.31.0
netCDF4==1.6.5
h5py==3.10.0
scipy==1.12.0
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
setuptools>=65.0.0
h5netcdf